Overhauls Modern Fortran rules.
Prior to the last commit, the dependence of the Modern Fortran rules on
the Punchcard Fortran rules placed restrictions on certain multi-line
Modern Fortran grammar rules. This commit takes advantage of this and
improves the robustness of several multi-line grammars. In addition,
naming and scope assignments are made more consistent and generally more
descriptive.

Due to the quantity of changes, care should be taken to verify that no
new bugs have been introduced and no old bugs have been reintroduced.