Fixes #37
Issue #37 was caused by including $self in the namelist rule set. This
allowed one of the other base language rules to force Atom into an
infinite loop of creating and then destroying type-specification-
statements blocks. To fix this the namelist rule set has been modified
to be more explicit as to which base rule sets are imported into it.