Fixes #55
Comments in free formed Fortran code should always start with !, while
comments in fixed form code may start with ! as well as c, C, and *
provided the last three are in the first character position of a given
line. The changes in this commit remove the options of c, C, and * from
the Modern Fortran grammar.