updates name-lists and operator interfaces
After the rule for variables was introduced variables would be matched
everywhere except in name-lists in type definition statements. This was
an oversight and has been fixed.

The scope for the operator keyword in operator interface constructs was
not correct in the interface construct end statement. This has been
fixed.