Fixes #67
One of the recent updates to Atom changed how the \\G tag was being used
in grammar regex rules. This caused certain attributes to be flagged as
invalid when they shouldn't have been. This commit fixes this issue by
removing the \\G from these rules. Unfortunately this also allows for
certain syntax mistakes to no longer be marked as incorrect.