Adds new functions to address #78
Toggling comments in Fortran - Fixed Form files would place the comment
character before the first non-blank character of the line. This is not
helpful as Fortran - Fixed Form requires comment characters to be placed
in the first character position of a given line. This commit does not
fix this issue but it adds functions that will be used to fix it in a
future commit.

The remaining step after this is to figure out how to make it so the
toggle comment keybinding is only overridden in Fortran - Fixed Form
files.