Fixes #92
Read and write statements were allowing type declaration statements to
appear after them on the same line which caused issues with certain
functions such as `real`, etc. To fix this a new IO-item-list rule has
been introduced and added to the rules for several IO-statements. This
was done broadly and might lead to some unintended highlighting for some
of the IO-statements and may need to be adjusted in the future.